debug ffmpeg

+++ ffmpeg.py
@@ -47,17 +47,23 @@ async def add_image_to_video(input_video, input_image, output_video, directory, 
     cmd = f"ffmpeg -loop 1 -i {input_image} -c:v {video_info['video_codec']} -t {image_duration} -vf scale={video_info['width']}:{video_info['height']} -r {video_info['framerate']} -b:v {video_info['bitrate']} -pix_fmt yuv420p {temp_video}"
     process = await asyncio.create_subprocess_shell(cmd)
     stdout, stderr = await process.communicate()
+    if process.returncode != 0:
+        raise Exception(f'FFMPEG error temp video: {stderr.decode()} \n{stdout.decode()}')
 
     # Convert the input video to MPEG-TS format
     input_video_ts = f"{directory}/input_video.ts"
     cmd = f"ffmpeg -i {input_video} -c copy -bsf:v h264_mp4toannexb -f mpegts {input_video_ts}"
     process = await asyncio.create_subprocess_shell(cmd)
     await process.communicate()
+    if process.returncode != 0:
+        raise Exception(f'FFMPEG error convert mpeg: {stderr.decode()} \n{stdout.decode()}')
 
     # Concatenate the original video and the temporary video
     cmd = f'ffmpeg -i "concat:{input_video_ts}|{temp_video}" -c copy -bsf:a aac_adtstoasc {output_video}'
     process = await asyncio.create_subprocess_shell(cmd)
     await process.communicate()
+    if process.returncode != 0:
+        raise Exception(f'FFMPEG error concat: {stderr.decode()} \n{stdout.decode()}')
 
     # Remove the temporary video and input_video_ts
     os.remove(temp_video)
